+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Unreleased
+
+* Fixes bug that not more than one custom token could be created at a time.
+
 ## 3.7.0 - 2017-12-08
 
 * Enable ordering by nested childs ([#135](https://github.com/kreait/firebase-php/pull/135))